backend/pyproject.toml
[project]
name = "voxharness-backend"
version = "0.1.0"
description = "Voice-first AI harness backend"
requires-python = ">=3.10"
dependencies = [
    "websockets>=13.0",
    "faster-whisper>=1.0",
    "silero-vad>=5.1",
    "anthropic>=0.40",
    "openai>=1.50",
    "elevenlabs>=1.0",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "soundfile>=0.12",
]

[project.scripts]
voxharness = "main:main"

[tool.setuptools.packages.find]
include = ["llm*"]

[tool.setuptools]
py-modules = ["main", "server", "config", "vad", "stt", "tts", "conversation", "tools"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
